# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
        "/token/create": {
            "get": {
                "operationId": "user.token.create",
                "summary": "Create a new token for the authenticated user",
                "tags": [
                    "AuthApi"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "device_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "token"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                },
                "security": []
            }
        },
        "/menus": {
            "get": {
                "operationId": "menus",
                "summary": "Get all menus",
                "tags": [
                    "BrowseMenuApi"
                ],
                "parameters": [
                    {
                        "name": "menu_id",
                        "in": "query",
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ]
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "string"
                                        },
                                        {
                                            "$ref": "#/components/schemas/MenuResource"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/menus/with-modifiers": {
            "get": {
                "operationId": "menus.with.modifiers",
                "summary": "Get menu with modifiers [Set Meal]",
                "tags": [
                    "BrowseMenuApi"
                ],
                "parameters": [
                    {
                        "name": "menu_id",
                        "in": "query",
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ]
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `MenuResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MenuResource"
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/MenuResource"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/menus/modifier-groups": {
            "get": {
                "operationId": "menus.modifier-groups",
                "description": "Returns all modifier groups. Adding a query param of modifiers = 1 will include the modifiers in the response",
                "summary": "Get all modifier groups",
                "tags": [
                    "BrowseMenuApi"
                ],
                "parameters": [
                    {
                        "name": "modifiers",
                        "in": "query",
                        "schema": {
                            "type": [
                                "boolean",
                                "null"
                            ]
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/menus/course": {
            "get": {
                "operationId": "menus.by.course",
                "summary": "Get all menus for the given course",
                "tags": [
                    "BrowseMenuApi"
                ],
                "parameters": [
                    {
                        "name": "course",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "starter"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `MenuResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MenuResource"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/menus/group": {
            "get": {
                "operationId": "menus.by.group",
                "summary": "Get all menus for the given group",
                "tags": [
                    "BrowseMenuApi"
                ],
                "parameters": [
                    {
                        "name": "group",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Sides"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `MenuResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MenuResource"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/menus/group-raw": {
            "get": {
                "operationId": "menus.group.raw",
                "description": "Query params:\n- group: string (optional) POS group name, defaults to 'Meat Order'\n\nThis returns the raw rows the stored procedure emits. Useful for clients\nthat need the exact POS-provided structure (menuName, groupName, computed_modifiers, etc.).",
                "summary": "Return raw stored-proc output for a POS group (calls `get_menus_by_group`)",
                "tags": [
                    "BrowseMenuApi"
                ],
                "parameters": [
                    {
                        "name": "group",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "Meat Order"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "enum": [
                                                "Failed to fetch group rows"
                                            ]
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "error"
                                    ]
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "rows": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "minItems": 0,
                                                    "maxItems": 0,
                                                    "additionalItems": false
                                                },
                                                "ids": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "minItems": 0,
                                                    "maxItems": 0,
                                                    "additionalItems": false
                                                },
                                                "packages": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "minItems": 0,
                                                    "maxItems": 0,
                                                    "additionalItems": false
                                                }
                                            },
                                            "required": [
                                                "rows",
                                                "ids",
                                                "packages"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "rows": {
                                                    "type": "string"
                                                },
                                                "ids": {
                                                    "type": "string"
                                                },
                                                "packages": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "rows",
                                                "ids",
                                                "packages"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/menus/modifiers-by-group": {
            "get": {
                "operationId": "menus.modifiers.by.grouped",
                "description": "Query params:\n- group: string (optional) POS group name, defaults to 'Meat Order'\n\nResponse: object with keys 'Pork','Beef','Chicken' mapping to arrays of modifiers.",
                "summary": "Return modifiers grouped by meat type (Pork/Beef/Chicken) for a POS menu group",
                "tags": [
                    "BrowseMenuApi"
                ],
                "parameters": [
                    {
                        "name": "group",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "Meat Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    },
                                    "minItems": 0,
                                    "maxItems": 0,
                                    "additionalItems": false
                                }
                            }
                        }
                    }
                }
            }
        },
        "/menus/package-modifiers": {
            "get": {
                "operationId": "menus.package.modifiers",
                "description": "Accepts a single parameter `package_id` (integer). For packages 46,47,48\nthis will return the exact modifiers selected by `Menu::getModifiers()`\nwhich uses `receipt_name` codes (P*, B*, C*).",
                "summary": "Return exact modifiers for a package menu by package id",
                "tags": [
                    "BrowseMenuApi"
                ],
                "parameters": [
                    {
                        "name": "package_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `MenuModifierResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MenuModifierResource"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/menus/category": {
            "get": {
                "operationId": "menus.by.category",
                "summary": "Get all menus for the given category",
                "tags": [
                    "BrowseMenuApi"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "beverage"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `MenuResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MenuResource"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/devices": {
            "get": {
                "operationId": "devices.index",
                "summary": "Returns a list of all devices",
                "tags": [
                    "DeviceApi"
                ],
                "responses": {
                    "200": {
                        "description": "Array of `DeviceResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DeviceResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "post": {
                "operationId": "devices.store",
                "summary": "Store a newly created device in storage",
                "tags": [
                    "DeviceApi"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreDeviceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "`DeviceResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/devices/refresh": {
            "post": {
                "operationId": "devices.refresh",
                "summary": "Revoke the current token and issue a new one with the same\nabilities and expiration time",
                "tags": [
                    "DeviceAuthApi"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "token": {
                                            "type": "string"
                                        },
                                        "device": {
                                            "type": "string"
                                        },
                                        "table": {
                                            "type": "string"
                                        },
                                        "expires_at": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "token",
                                        "device",
                                        "table",
                                        "expires_at"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/tables/services": {
            "get": {
                "operationId": "tables.services",
                "summary": "Returns a list of all services",
                "tags": [
                    "TableServiceApi"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
...
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}